+++ examples/simple-fxaa.rs
@@ -55,7 +55,7 @@ fn main() {
     
     let mut frame_start = Instant::now();
 
-    thin_engine::builder(input).with_setup(|display, window| {
+    thin_engine::builder(input).with_setup(|display, window, _| {
         window.set_title("FXAA Test");
         let _ = window.set_cursor_grab(CursorGrabMode::Confined);
         let _ = window.set_cursor_grab(CursorGrabMode::Locked);